fixed preemption for will run calls

+++ trunk/src/common/node_select.c
@@ -82,7 +82,7 @@ typedef struct slurm_select_ops {
 						uint32_t min_nodes, 
 						uint32_t max_nodes,
 						uint32_t req_nodes,
-						int mode,
+						uint16_t mode,
 						List preeemptee_candidates,
 						List *preemptee_job_list);
 	int		(*job_begin)	       (struct job_record *job_ptr);
@@ -642,7 +642,7 @@ extern int select_g_block_init(List block_list)
  */
 extern int select_g_job_test(struct job_record *job_ptr, bitstr_t *bitmap,
 			     uint32_t min_nodes, uint32_t max_nodes, 
-			     uint32_t req_nodes, int mode, 
+			     uint32_t req_nodes, uint16_t mode, 
 			     List preemptee_candidates,
 			     List *preemptee_job_list)
 {
